Extract TaskQueue, remove childProcess property from Task, add doc
